add: avances de proyecto final con estilos css

+++ Proyecto-Final/borrar.php
@@ -1,42 +1,18 @@
-﻿<?php
-  // 1) Conexion
-  // a) realizar la conexion con la bbdd
-  // b) seleccionar la base de datos a usar
-
-  // 2) Almacenamos los datos del envío GET
-  // a) generar variables para el id a utilizar
-
-  // 3) Preparar la orden SQL
-  // DELETE FROM nombre_tabla WHERE campo_tabla=dato
-  // => Elimina de la siguiente tabla el registro donde este campo sea igual a este dato
-  // DELETE FROM nombre_tabla
-  // => Elimina todos los registros de la siguiente tabla
-  // a) generar la consulta a realizar
-
-  // 4) Ejecutar la orden y eliminamos el regitro seleccionado
-  // a) ejecutar la consulta
-  // a) rederigir a index
-?>
 <?php
 // 1) Conexion
-// a) realizar la conexion con la bbdd
-// b) seleccionar la base de datos a usar
 $conexion = mysqli_connect("127.0.0.1", "root", "");
 mysqli_select_db($conexion, "tienda");
 
-// 2) Almacenamos los datos del envío GET
-// a) generar variables para el id a utilizar
-$id = $_GET["id"];
+// 2) Almacenamos el ID recibido por GET
+$id = isset($_GET["id"]) ? (int) $_GET["id"] : 0;
 
-// 3) Preparar la orden SQL
-// DELETE FROM nombre_tabla WHERE campo_tabla=dato
-// => Elimina de la siguiente tabla el registro donde este campo sea igual a este dato
-// a) generar la consulta a realizar
-$consulta = "DELETE FROM ropa WHERE id=$id";
+// 3) Eliminamos el registro solo si el ID es valido
+if ($id > 0) {
+    $consulta = "DELETE FROM ropa WHERE id=$id";
+    mysqli_query($conexion, $consulta);
+}
 
-// 4) Ejecutar la orden y eliminamos el regitro seleccionado
-// a) ejecutar la consulta
-// a) rederigir a index
-mysqli_query($conexion, $consulta);
-
+// 4) Redirigir al listado
 header("location:listar.php");
+exit;
+?>
